Get questions to show up again in the authoring tool
Some mysterious force was acting such that when the watch statement first
executes, $scope.init is the init function in the vgQuestions directive, but
when it executes again, it actually refers to the $scope.init function in the
vgQuestionMultiple directive (this is probably dependant on what types of
questions are in play). A local reference is now stored such that the function
can be called reliably.